Analyzer of an artificial Neural Network

This class analyzes a passed neural network and stores its internal structures and layers independently of the actual class of the network. With the help of this object, various methods of interpretable machine learning can be applied to it for a better understanding of individual predictions or of the whole model. You can use models from the following libraries:

keras,

neuralnet

Implemented methods

Methods that aim to explain an individual prediction of a neural network are called local. In contrast, global methods provide an explanation for the entire model. An object of the Analyzer class can be applied to the following local and global methods:

Global:

Connection_Weights, Olden et al. (2004)

Local:

Layerwise Relevance Propagation (LRP), Bach et al. (2015)

Deep Learning Important Feartures (DeepLift), Shrikumar et al. (2017)

SmoothGrad, Smilkov et al. (2017)

Gradient

References

J. D. Olden et al. (2004) An accurate comparison of methods for quantifying variable importance in artificial neural networks using simulated data. Ecological Modelling 178, p. 389–397

S. Bach et al. (2015) On pixel-wise explanations for non-linear classifier decisions by layer-wise relevance propagation. PLoS ONE 10, p. 1-46

A. Shrikumar et al. (2017) Learning important features through propagating activation differences. ICML 2017, p. 4844-4866

D. Smilkov et al. (2017) SmoothGrad: removing noise by adding noise. CoRR, abs/1706.03825

Public fields

model

The given neural network.

layers

List of layers in the model. It contains only layers that are relevant for the evaluation. For example a list of Dense_Layer.

num_layers

Number of all layers in layers.

last_input

Last recorded input for the forward pass (default: NULL).

last_input_ref

Last recorded reference input for the forward pass (default: NULL).

dim_in

Dimension of the input features.

dim_out

Dimension of the models output, i.e. dimension of the response variables.

feature_names

A list of names for the input features.

response_names

A list of names for the response variables.

Methods

Public methods

Analyzer$new()

Analyzer$forward()

Analyzer$update()

Analyzer$clone()

Method new()

Create a new analyzer for a given neural network.

Usage

Analyzer$new(model, feature_names = NULL, response_names = NULL)

Arguments

model

A trained neural network for classification or regression tasks to be interpreted. Only models from the following types or packages are allowed: keras_model, keras_model_sequential or neuralnet.

feature_names

A list of names for the input features. Use the default value NULL for the default names (X1, X2, ...).

response_names

A list of names for the response variables. Use the default value NULL for the default names (Y1, Y2, ...).

Returns

A new instance of the R6 class 'Analyzer'.

Method forward()

The forward method of the whole model, i.e. it calculates the output \(y=f(x)\) of a given input \(x\) respectively an reference input \(x'\). In doing so all intermediate values are stored in the individual layers. A batch-wise evaluation is performed, hence \(x\) must be a matrix of inputs.

Usage

Analyzer$forward(x, x_ref = NULL)

Arguments

x

Input matrix of the model with size (num_data, dim_in).

x_ref

Reference input vector of the model. If this value is not needed, it can be set to the default value NULL.

Returns

A list with two vectors. The first one is the output for the inputs x and the second entry is the output for the reference input x_ref.

Method update()

This method updates the stored intermediate values in each layer from the list layers when the inputs x or reference input x_ref has changed.

Usage

Analyzer$update(x, x_ref = NULL)

Arguments

x

Input matrix of the model with size (num_data, dim_in).

x_ref

Reference input vector of the model. If this value is not needed, it can be set to the default value NULL.

Returns

Returns the instance itself.

Method clone()

The objects of this class are cloneable with this method.

Usage

Analyzer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

#------------------------- neuralnet model ----------------------------------
library(neuralnet)
nn <- neuralnet((Species == "setosa") ~ Petal.Length + Petal.Width,
                iris, linear.output = FALSE,
hidden = c(3,2), act.fct = "tanh", rep = 1)
analyzer = Analyzer$new(nn)

#-------------------------- keras model -------------------------------------
library(keras)

iris[,5] <- as.numeric(iris[,5]) -1
# Turn `iris` into a matrix
iris <- as.matrix(iris)
# Set iris `dimnames` to `NULL`
dimnames(iris) <- NULL
iris.training <- iris[, 1:4]
iris.trainingtarget <- iris[, 5]
# One hot encode training target values
iris.trainLabels <- to_categorical(iris.trainingtarget)

#> Error: Error 1 occurred creating conda environment r-reticulate

# Define model
model <- keras_model_sequential()

#> Error: Error 1 occurred creating conda environment r-reticulate

model %>%
  layer_dense(units = 16, activation = 'relu', input_shape = c(4)) %>%
  layer_dropout(0.1) %>%
  layer_dense(units = 8, activation = 'relu') %>%
  layer_dropout(0.1) %>%
  layer_dense(units = 3, activation = 'softmax')

#> Error: Error 1 occurred creating conda environment r-reticulate

# Compile the model
model %>% compile(
  loss = 'categorical_crossentropy',
  optimizer = 'adam',
  metrics = 'accuracy'
)

#> Error in compile(., loss = "categorical_crossentropy", optimizer = "adam",     metrics = "accuracy"): object 'model' not found

# Train the model
history <- model %>% fit(
  iris.training,
  iris.trainLabels,
  epochs = 50,
  batch_size = 5,
  validation_split = 0.2, verbose = 1
)

#> Error in fit(., iris.training, iris.trainLabels, epochs = 50, batch_size = 5,     validation_split = 0.2, verbose = 1): object 'model' not found

analyzer = Analyzer$new(model)

#> Error in inherits(model, "nn_module"): object 'model' not found