Dense layer of a Neural Network

Implementation of a regular densely-connected Neural Network layer as an R6 class, where input, preactivation and output values of the last forward pass are stored (same for a reference input, if this is needed). Applies a linear transformation followed by an activation function \(\sigma\) to the incoming data, i.e. $$y = \sigma (W^T x + b).$$

Public fields

type

Type of the layer (in this case "Dense").

dim

Dimension of this layer, i.e. (dim_in, dim_out).

weights

The weight matrix \(W\) of the dense layer with size (dim_in, dim_out).

bias

Bias vector \(b\) of the linear transformation. It's a vector of length dim_out.

activation

Activation function \(\sigma\) to turn the linear transformation into a non-linear one.

activation_name

Name of the activation function \(\sigma\).

inputs

Save the inputs from the last forward pass of this layer. If there was no call of method Dense_Layer$forward yet then this value is NULL.

preactivation

Save the outputs of the linear transformation from the last forward pass of this layer. If there was no call of method Dense_Layer$forward yet, then this value is NULL.

outputs

Save the outputs of the whole layer from the last forward pass. If there was no call of method Dense_Layer$forward yet, then this value is NULL.

inputs_ref

Save the reference inputs from the last forward pass of this layer. If there was no call of method Dense_Layer$forward yet, then this value is NULL.

preactivation_ref

Save the reference outputs of the linear transformation from the last forward pass of this layer. If there was no call of method Dense_Layer$forward yet, then this value is NULL.

outputs_ref

Save the reference outputs of the whole layer from the last forward pass. If there was no call of method Dense_Layer$forward yet, then this value is NULL.

Methods

Public methods

Dense_Layer$new()

Dense_Layer$forward()

Dense_Layer$clone()

Method new()

Create a new instance of this class with given parameters of a dense layer.

Usage

Dense_Layer$new(weights, bias, activation, activation_name = NULL)

Arguments

weights

The weight matrix of dimension (dim_in , dim_out) for the linear transformation.

bias

The bias vector of length dim_out for the linear transformation.

activation

The activation function of the dense layer.

activation_name

Name of the activation function.

Returns

A new instance of the R6 class 'Dense_Layer' with the given parameters.

Method forward()

The forward method of the dense layer. This method calculates the linear transformation with the following non-linearity for the input x and, if needed, the reference value x_ref. In addition, the input x must be a matrix of size (batch_size x dim_in), which enables batch-wise evaluation. All the intermediate values are stored in the class attributes: inputs, preactivation, outputs, inputs_ref, preactivation_ref, outputs_ref.

Usage

Dense_Layer$forward(x, x_ref = NULL)

Arguments

x

Input matrix of size (batch_size x dim_in) for this layer.

x_ref

Reference input vector for this layer. It is only needed for the DeepLift method and can otherwise be set to NULL.

Returns

A list of two elements. The first one (out) is the output matrix of this layer for the input x and the second entry (out_ref) is the output for the reference input x_ref.

Method clone()

The objects of this class are cloneable with this method.

Usage

Dense_Layer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Weight matrix of shape (number input features x number outputs)
# for example (3,5)
W <- matrix(c(1, -1, 2, 4, -1,
              2, -5, 3, 1, -2,
              3, 4, -1, -2, 4), nrow = 3, ncol = 5, byrow = TRUE)
# Bias vector of length (number outputs)
b <- c(-2, 1, 3, 4, -5)
# Activation function
sigmoid <- function(x) 1 / (1 + exp(-x))
# Create a new dense layer
my_layer <- Dense_Layer$new(weights = W, bias = b, activation = sigmoid)

# example forward pass of a batch of size 4 with no reference input
inputs <- matrix(rnorm(12), ncol = 3, nrow = 4)
my_layer$forward(x = inputs)

#> $out
#>          [,1]      [,2]      [,3]       [,4]        [,5]
#> B1 0.98343927 0.6418486 0.9990927 0.99986192 0.028080271
#> B2 0.98599527 0.0348894 0.9940901 0.09021834 0.048189364
#> B3 0.02838506 0.9932418 0.2499623 0.08188519 0.111237857
#> B4 0.01150540 0.9790393 0.8765943 0.99928567 0.004728206
#> 
#> $out_ref
#> NULL
#> 

# get preactivation of the last forward pass
my_layer$preactivation

#>         [,1]       [,2]      [,3]      [,4]      [,5]
#> B1  4.084022  0.5833968  7.004095  8.887529 -3.544206
#> B2  4.254257 -3.3200596  5.125207 -2.310972 -2.983228
#> B3 -3.533096  4.9902144 -1.098813 -2.417004 -2.078159
#> B4 -4.453367  3.8439227  1.960567  7.243456 -5.349470