Calculate the Gradient

This method computes the gradients of the outputs with respect to the input variables, i.e. for all input variable \(i\) and output class \(j\) $$d f(x)_j / d x_i.$$

Gradient(analyzer, data, times_input = FALSE, ignore_last_act = TRUE)

Arguments

analyzer
An instance of the R6 class Analyzer.

data
Either a matrix or a data frame, where each row must describe an input to the network.

times_input
Multiplies the gradients with the input features. This method is called 'Gradient x Inputs'.

ignore_last_act
Calculate the gradients of the output or the preactivation of the output (default: TRUE).

Value

Returns an array of size (dim_in, dim_out, num_data) which contains the gradients for each input variable to the output predictions for each element in the given data.

Examples

library(neuralnet)
nn <- neuralnet( Species ~ .,
                 iris, linear.output = FALSE,
                 hidden = c(10,8), act.fct = "tanh", rep = 1, threshold = 0.5)
# create an analyzer for this model
analyzer = Analyzer$new(nn)

# calculate gradients for the whole dataset
Gradient(analyzer, data = iris[,-5])

#> , , B1
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0002960778 -0.0003777141 -3.037775e-05
#> Sepal.Width  -0.0008156319 -0.0016171805 -1.063295e-03
#> Petal.Length  0.0009964913  0.0032124287  1.322071e-03
#> Petal.Width   0.0020871373  0.0047143738  2.140639e-03
#> 
#> , , B2
#> 
#>                     setosa   versicolor    virginica
#> Sepal.Length  0.0004155268 -0.000900464 -0.000139769
#> Sepal.Width  -0.0021127685 -0.003482180 -0.002586654
#> Petal.Length  0.0024767869  0.007126997  0.002899332
#> Petal.Width   0.0047570183  0.010279253  0.004686776
#> 
#> , , B3
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0005576538 -0.0004504674  1.255743e-05
#> Sepal.Width  -0.0011693815 -0.0021522848 -1.613103e-03
#> Petal.Length  0.0009630927  0.0043547926  1.517269e-03
#> Petal.Width   0.0025185960  0.0065244923  2.604162e-03
#> 
#> , , B4
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0007470782 -0.0008498264 -0.0002440318
#> Sepal.Width  -0.0022542251 -0.0046311989 -0.0027652799
#> Petal.Length  0.0032036292  0.0086847165  0.0042040987
#> Petal.Width   0.0064451396  0.0132480298  0.0064148633
#> 
#> , , B5
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0003417167 -0.0002943154 -1.440687e-05
#> Sepal.Width  -0.0006629507 -0.0014745510 -9.044514e-04
#> Petal.Length  0.0007994871  0.0028451432  1.189444e-03
#> Petal.Width   0.0018375314  0.0042849669  1.918103e-03
#> 
#> , , B6
#> 
#>                    setosa   versicolor     virginica
#> Sepal.Length -0.000127813 -0.001081411 -0.0004471787
#> Sepal.Width  -0.002259825 -0.004101301 -0.0023206399
#> Petal.Length  0.004338791  0.007854711  0.0044471553
#> Petal.Width   0.006774824  0.011210605  0.0064798563
#> 
#> , , B7
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0005639901 -0.0006524149 -0.0001591037
#> Sepal.Width  -0.0015335811 -0.0034203723 -0.0019072417
#> Petal.Length  0.0023714253  0.0063832037  0.0031348375
#> Petal.Width   0.0047281947  0.0096742118  0.0047792155
#> 
#> , , B8
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0003575164 -0.0005738713 -0.0001180217
#> Sepal.Width  -0.0012729754 -0.0025864531 -0.0015619763
#> Petal.Length  0.0018723863  0.0049847534  0.0023347007
#> Petal.Width   0.0035961431  0.0073725133  0.0036134139
#> 
#> , , B9
#> 
#>                    setosa    versicolor     virginica
#> Sepal.Length  0.001021067 -0.0008815348 -0.0001751793
#> Sepal.Width  -0.002560610 -0.0049739007 -0.0032859684
#> Petal.Length  0.002927311  0.0095140788  0.0041397179
#> Petal.Width   0.006523956  0.0145971938  0.0065277923
#> 
#> , , B10
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0005512544 -0.0005625981 -6.207992e-05
#> Sepal.Width  -0.0015195781 -0.0027204858 -1.969285e-03
#> Petal.Length  0.0015449182  0.0054098511  2.131818e-03
#> Petal.Width   0.0034753668  0.0081163896  3.484550e-03
#> 
#> , , B11
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0001769978 -0.0003563199 -5.981944e-05
#> Sepal.Width  -0.0007169851 -0.0014422022 -8.795212e-04
#> Petal.Length  0.0010747307  0.0028296785  1.296961e-03
#> Petal.Width   0.0020042929  0.0041028161  2.025313e-03
#> 
#> , , B12
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0005524047 -0.0006426909 -0.0002372286
#> Sepal.Width  -0.0017209985 -0.0037352698 -0.0020565612
#> Petal.Length  0.0027471079  0.0068377511  0.0035596006
#> Petal.Width   0.0053598516  0.0105241226  0.0053154163
#> 
#> , , B13
#> 
#>                     setosa   versicolor     virginica
#> Sepal.Length  0.0006266734 -0.000513160  0.0000188802
#> Sepal.Width  -0.0015558027 -0.002389237 -0.0021076554
#> Petal.Length  0.0009639869  0.005021738  0.0015178256
#> Petal.Width   0.0027309742  0.007509800  0.0027403742
#> 
#> , , B14
#> 
#>                     setosa   versicolor     virginica
#> Sepal.Length  9.299424e-04 -0.000051478  2.391162e-04
#> Sepal.Width  -1.232400e-03 -0.001029205 -1.990222e-03
#> Petal.Length -1.135596e-03  0.002696924 -4.874164e-04
#> Petal.Width   5.849834e-05  0.004483534  6.095297e-05
#> 
#> , , B15
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  9.238725e-05 -0.0000961772  1.877127e-05
#> Sepal.Width  -2.246377e-04 -0.0003204390 -3.205403e-04
#> Petal.Length  9.362382e-05  0.0007360557  1.566680e-04
#> Petal.Width   3.166747e-04  0.0010400035  3.354443e-04
#> 
#> , , B16
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  4.490888e-05 -0.0002744072 -7.606276e-05
#> Sepal.Width  -5.087729e-04 -0.0010463721 -5.691244e-04
#> Petal.Length  9.600886e-04  0.0020135958  1.057623e-03
#> Petal.Width   1.594968e-03  0.0028771562  1.583075e-03
#> 
#> , , B17
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  9.363359e-05 -0.0003699181 -7.734331e-05
#> Sepal.Width  -6.896037e-04 -0.0013641370 -8.053726e-04
#> Petal.Length  1.160155e-03  0.0026868723  1.300165e-03
#> Petal.Width   1.994767e-03  0.0038189060  1.996011e-03
#> 
#> , , B18
#> 
#>                     setosa   versicolor     virginica
#> Sepal.Length  0.0002122258 -0.000616036 -0.0001317388
#> Sepal.Width  -0.0012362008 -0.002409696 -0.0014620029
#> Petal.Length  0.0019568647  0.004727252  0.0022516842
#> Petal.Width   0.0034756936  0.006805831  0.0034690017
#> 
#> , , B19
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length -6.886895e-05 -0.0007732307 -0.0002916709
#> Sepal.Width  -1.582090e-03 -0.0028527386 -0.0016505577
#> Petal.Length  2.962910e-03  0.0055188011  0.0030384617
#> Petal.Width   4.651170e-03  0.0078342043  0.0044659900
#> 
#> , , B20
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0002288814 -0.0004744957 -0.0001213703
#> Sepal.Width  -0.0009697235 -0.0021067398 -0.0011491102
#> Petal.Length  0.0016808337  0.0039864409  0.0020254087
#> Petal.Width   0.0030481329  0.0058716498  0.0030608177
#> 
#> , , B21
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  4.945162e-05 -0.0009934188 -0.0003433107
#> Sepal.Width  -2.097787e-03 -0.0038271199 -0.0022722021
#> Petal.Length  3.688612e-03  0.0074144355  0.0039180962
#> Petal.Width   6.023897e-03  0.0106305701  0.0058274653
#> 
#> , , B22
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  7.329014e-05 -0.0009114505 -0.0003075779
#> Sepal.Width  -1.833487e-03 -0.0035487207 -0.0019907155
#> Petal.Length  3.400805e-03  0.0068085863  0.0036633429
#> Petal.Width   5.561566e-03  0.0097741652  0.0054287686
#> 
#> , , B23
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0003921330 -5.675596e-05  1.098155e-04
#> Sepal.Width  -0.0003367728 -4.846449e-04 -6.295724e-04
#> Petal.Length -0.0003720873  1.140966e-03 -5.686604e-05
#> Petal.Width   0.0001074768  1.827192e-03  2.027987e-04
#> 
#> , , B24
#> 
#>                    setosa  versicolor    virginica
#> Sepal.Length -0.001371139 -0.00421609 -0.002011517
#> Sepal.Width  -0.009103853 -0.01519558 -0.008934451
#> Petal.Length  0.017569766  0.02931979  0.017132398
#> Petal.Width   0.026233636  0.04139165  0.024642037
#> 
#> , , B25
#> 
#>                     setosa   versicolor     virginica
#> Sepal.Length  0.0005804396 -0.001315882 -0.0009489572
#> Sepal.Width  -0.0043048396 -0.008465529 -0.0045731347
#> Petal.Length  0.0076180193  0.015046844  0.0089925015
#> Petal.Width   0.0134965184  0.023492255  0.0128045211
#> 
#> , , B26
#> 
#>                     setosa   versicolor     virginica
#> Sepal.Length  0.0002120784 -0.001509664 -0.0004826024
#> Sepal.Width  -0.0033356282 -0.005923395 -0.0037141638
#> Petal.Length  0.0053866753  0.011578258  0.0058354510
#> Petal.Width   0.0090759144  0.016681819  0.0087923129
#> 
#> , , B27
#> 
#>                     setosa   versicolor     virginica
#> Sepal.Length -0.0001839891 -0.001927856 -0.0007824233
#> Sepal.Width  -0.0040550732 -0.007327455 -0.0042043074
#> Petal.Length  0.0076346794  0.014071453  0.0078670496
#> Petal.Width   0.0120002170  0.020098781  0.0114964062
#> 
#> , , B28
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0002479451 -0.0004983514 -9.510077e-05
#> Sepal.Width  -0.0010478329 -0.0020690764 -1.274121e-03
#> Petal.Length  0.0015544505  0.0040454895  1.871283e-03
#> Petal.Width   0.0028988148  0.0058951619  2.909727e-03
#> 
#> , , B29
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0002529483 -0.0004569725 -5.019055e-05
#> Sepal.Width  -0.0009916903 -0.0017737161 -1.247118e-03
#> Petal.Length  0.0012003291  0.0036050289  1.463917e-03
#> Petal.Width   0.0023612751  0.0051986558  2.376635e-03
#> 
#> , , B30
#> 
#>                     setosa   versicolor     virginica
#> Sepal.Length  0.0006423301 -0.000941747 -0.0003576434
#> Sepal.Width  -0.0024751520 -0.005122154 -0.0028995454
#> Petal.Length  0.0039522984  0.009468618  0.0049343056
#> Petal.Width   0.0074751676  0.014434153  0.0073563692
#> 
#> , , B31
#> 
#>                     setosa   versicolor    virginica
#> Sepal.Length  0.0004709626 -0.001235221 -0.000415943
#> Sepal.Width  -0.0029039023 -0.005622994 -0.003323754
#> Petal.Length  0.0046986197  0.010684597  0.005471859
#> Petal.Width   0.0083808006  0.015822982  0.008205664
#> 
#> , , B32
#> 
#>                     setosa   versicolor     virginica
#> Sepal.Length -0.0001944573 -0.001332469 -0.0005068301
#> Sepal.Width  -0.0027681581 -0.004753315 -0.0028694948
#> Petal.Length  0.0050432410  0.009296669  0.0050640355
#> Petal.Width   0.0078210144  0.013109269  0.0074571304
#> 
#> , , B33
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0002702060 -2.707502e-05  1.461383e-05
#> Sepal.Width  -0.0001819744 -6.131899e-04 -3.167618e-04
#> Petal.Length  0.0001244126  1.054915e-03  4.432312e-04
#> Petal.Width   0.0006475103  1.806444e-03  7.143897e-04
#> 
#> , , B34
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0001385760 -0.0001100904  7.253376e-06
#> Sepal.Width  -0.0002007888 -0.0005196107 -2.966493e-04
#> Petal.Length  0.0002491323  0.0010058946  4.037179e-04
#> Petal.Width   0.0006098647  0.0014991252  6.646612e-04
#> 
#> , , B35
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0003996956 -0.0009788797 -0.0002337094
#> Sepal.Width  -0.0022073729 -0.0040676656 -0.0026143494
#> Petal.Length  0.0031983071  0.0079803128  0.0037293280
#> Petal.Width   0.0058396958  0.0116373352  0.0057619253
#> 
#> , , B36
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0003549489 -0.0003848327  5.097757e-05
#> Sepal.Width  -0.0010681838 -0.0013511956 -1.459292e-03
#> Petal.Length  0.0004414036  0.0031181454  6.645701e-04
#> Petal.Width   0.0014080315  0.0044600000  1.408576e-03
#> 
#> , , B37
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0001634622 -0.0002905647 -0.0000047627
#> Sepal.Width  -0.0006860568 -0.0009964558 -0.0008848543
#> Petal.Length  0.0005713342  0.0021782243  0.0006875933
#> Petal.Width   0.0012299510  0.0030814916  0.0012238962
#> 
#> , , B38
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0004428507 -9.163205e-05  3.805959e-05
#> Sepal.Width  -0.0004249456 -1.027386e-03 -6.870026e-04
#> Petal.Length  0.0001707885  1.913358e-03  6.484197e-04
#> Petal.Width   0.0010330379  3.146946e-03  1.126948e-03
#> 
#> , , B39
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0009311554 -0.0005349889 -1.077567e-06
#> Sepal.Width  -0.0017014579 -0.0032316994 -2.362446e-03
#> Petal.Length  0.0013207459  0.0063314311  2.301946e-03
#> Petal.Width   0.0037766167  0.0098087132  3.873132e-03
#> 
#> , , B40
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0002940839 -0.0005882002 -0.0001195111
#> Sepal.Width  -0.0012659338 -0.0024775468 -0.0015329678
#> Petal.Length  0.0018674651  0.0048346074  0.0022468074
#> Petal.Width   0.0034843968  0.0070647006  0.0034859065
#> 
#> , , B41
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0002719358 -0.0004744351 -5.599237e-05
#> Sepal.Width  -0.0009568131 -0.0018884069 -1.207110e-03
#> Petal.Length  0.0013089671  0.0037578284  1.620633e-03
#> Petal.Width   0.0025361199  0.0054361040  2.583219e-03
#> 
#> , , B42
#> 
#>                     setosa  versicolor     virginica
#> Sepal.Length  0.0003662531 -0.00339829 -0.0008366333
#> Sepal.Width  -0.0084945474 -0.01178690 -0.0097755992
#> Petal.Length  0.0099638086  0.02477742  0.0102091060
#> Petal.Width   0.0172937636  0.03500432  0.0163173749
#> 
#> , , B43
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0008830606 -0.0002672063  2.014597e-05
#> Sepal.Width  -0.0011094972 -0.0024763392 -1.640242e-03
#> Petal.Length  0.0007678704  0.0046108260  1.751184e-03
#> Petal.Width   0.0027702847  0.0074758111  2.902810e-03
#> 
#> , , B44
#> 
#>                    setosa   versicolor    virginica
#> Sepal.Length -0.001182445 -0.003720664 -0.001759737
#> Sepal.Width  -0.007983182 -0.013408922 -0.007837425
#> Petal.Length  0.015484323  0.025854921  0.015119450
#> Petal.Width   0.023139484  0.036496217  0.021754720
#> 
#> , , B45
#> 
#>                     setosa   versicolor    virginica
#> Sepal.Length -0.0003958885 -0.002216519 -0.001136416
#> Sepal.Width  -0.0051470024 -0.009083441 -0.005148578
#> Petal.Length  0.0098623211  0.017114989  0.010098723
#> Petal.Width   0.0153635421  0.024817455  0.014488148
#> 
#> , , B46
#> 
#>                     setosa   versicolor     virginica
#> Sepal.Length  0.0002878192 -0.001424098 -0.0003776183
#> Sepal.Width  -0.0030660303 -0.005424039 -0.0035046345
#> Petal.Length  0.0047050189  0.010744513  0.0051575048
#> Petal.Width   0.0080851741  0.015398694  0.0078979660
#> 
#> , , B47
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0003232793 -0.0003311659 -9.485349e-05
#> Sepal.Width  -0.0008124907 -0.0018947018 -1.006102e-03
#> Petal.Length  0.0013166789  0.0034700299  1.768373e-03
#> Petal.Width   0.0026410740  0.0053270458  2.666839e-03
#> 
#> , , B48
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0007019409 -0.0005205035 -7.239426e-05
#> Sepal.Width  -0.0014380479 -0.0030359757 -1.907575e-03
#> Petal.Length  0.0016646557  0.0057569559  2.493310e-03
#> Petal.Width   0.0038758443  0.0088702981  3.952411e-03
#> 
#> , , B49
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0002126362 -0.0003524987 -5.733343e-05
#> Sepal.Width  -0.0007197309 -0.0015052948 -8.960392e-04
#> Petal.Length  0.0010783543  0.0029226127  1.345908e-03
#> Petal.Width   0.0020659712  0.0042797155  2.099297e-03
#> 
#> , , B50
#> 
#>                     setosa    versicolor     virginica
#> Sepal.Length  0.0003544198 -0.0005357364 -6.041486e-05
#> Sepal.Width  -0.0011993922 -0.0022192410 -1.523214e-03
#> Petal.Length  0.0014476586  0.0044505345  1.831533e-03
#> Petal.Width   0.0029301550  0.0064956912  2.959164e-03
#> 
#> , , B51
#> 
#>                    setosa versicolor   virginica
#> Sepal.Length -0.006414214  0.2648768  0.01009305
#> Sepal.Width  -0.007705724  0.2937509  0.01435335
#> Petal.Length  0.011645741 -0.5013681 -0.02107154
#> Petal.Width   0.012182102 -0.4829779 -0.02482427
#> 
#> , , B52
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.01951416  0.3750545  0.005422393
#> Sepal.Width  -0.02516073  0.4294024  0.013424321
#> Petal.Length  0.03838002 -0.7198657 -0.014847419
#> Petal.Width   0.04211187 -0.7125315 -0.024919689
#> 
#> , , B53
#> 
#>                   setosa versicolor   virginica
#> Sepal.Length -0.03245967  0.4898221  0.00417962
#> Sepal.Width  -0.04011240  0.5552926  0.01073695
#> Petal.Length  0.06414270 -0.9412011 -0.01174695
#> Petal.Width   0.06766291 -0.9230969 -0.02004918
#> 
#> , , B54
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.02283211  0.4854913 -0.006417344
#> Sepal.Width  -0.03940495  0.5960001  0.011224236
#> Petal.Length  0.05049625 -0.9544995  0.001055696
#> Petal.Width   0.06900883 -1.0020033 -0.025299363
#> 
#> , , B55
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.03304800  0.5211478  0.003130721
#> Sepal.Width  -0.04126641  0.5909028  0.009808426
#> Petal.Length  0.06547268 -1.0010586 -0.009870487
#> Petal.Width   0.06967497 -0.9820021 -0.018583066
#> 
#> , , B56
#> 
#>                    setosa versicolor   virginica
#> Sepal.Length  0.009282942  0.3793998 -0.03996268
#> Sepal.Width  -0.036606285  0.6364857  0.01616959
#> Petal.Length  0.010569809 -0.8546597  0.03870464
#> Petal.Width   0.077487053 -1.1319450 -0.04943404
#> 
#> , , B57
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.02638461  0.4964517 -0.003597393
#> Sepal.Width  -0.04125641  0.5982881  0.010892323
#> Petal.Length  0.05737776 -0.9757035 -0.002361512
#> Petal.Width   0.07245545 -1.0074076 -0.023598941
#> 
#> , , B58
#> 
#>                    setosa  versicolor    virginica
#> Sepal.Length  0.008698735  0.12730370 -0.007680098
#> Sepal.Width   0.020209423  0.10827014  0.009609689
#> Petal.Length -0.047864585 -0.10808479  0.007987411
#> Petal.Width  -0.057931834 -0.07204809 -0.018934032
#> 
#> , , B59
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.01761801  0.3604447  0.005433348
#> Sepal.Width  -0.02295601  0.4133068  0.013684752
#> Petal.Length  0.03454119 -0.6913231 -0.014998542
#> Petal.Width   0.03829271 -0.6853170 -0.025424483
#> 
#> , , B60
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.01537672  0.4226037 -0.014653209
#> Sepal.Width  -0.03812592  0.5649695  0.013589870
#> Petal.Length  0.03984111 -0.8551611  0.009676729
#> Petal.Width   0.06887436 -0.9623285 -0.033482986
#> 
#> , , B61
#> 
#>                    setosa versicolor    virginica
#> Sepal.Length -0.005217918  0.2906101 -0.009156397
#> Sepal.Width  -0.011969483  0.3610013  0.014579298
#> Petal.Length  0.002608353 -0.5289007  0.003478391
#> Petal.Width   0.012830666 -0.5710646 -0.032336164
#> 
#> , , B62
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.02298418  0.4117856  0.002064688
#> Sepal.Width  -0.03133753  0.4807415  0.012827856
#> Petal.Length  0.04595669 -0.7944072 -0.010338706
#> Petal.Width   0.05280393 -0.7997366 -0.025068552
#> 
#> , , B63
#> 
#>                    setosa versicolor    virginica
#> Sepal.Length -0.005589625  0.2625284  0.005326381
#> Sepal.Width  -0.006308314  0.2929168  0.013999035
#> Petal.Length  0.005669409 -0.4791540 -0.014378190
#> Petal.Width   0.006276815 -0.4666549 -0.025555310
#> 
#> , , B64
#> 
#>                    setosa versicolor   virginica
#> Sepal.Length -0.004818657  0.4586721 -0.01756941
#> Sepal.Width  -0.032016574  0.6218625  0.01228279
#> Petal.Length  0.025355696 -0.9442924  0.01380621
#> Petal.Width   0.062671044 -1.0709961 -0.03223652
#> 
#> , , B65
#> 
#>                    setosa versicolor    virginica
#> Sepal.Length  0.005654126  0.1415457  0.006274412
#> Sepal.Width   0.010675829  0.1370884  0.010996466
#> Petal.Length -0.021708176 -0.2177763 -0.012294856
#> Petal.Width  -0.026118251 -0.1863592 -0.017910967
#> 
#> , , B66
#> 
#>                    setosa versicolor   virginica
#> Sepal.Length -0.001811074  0.2244284  0.01039835
#> Sepal.Width  -0.001891640  0.2459045  0.01413544
#> Petal.Length  0.001832491 -0.4193948 -0.02115089
#> Petal.Width   0.001801165 -0.4002061 -0.02412443
#> 
#> , , B67
#> 
#>                   setosa versicolor   virginica
#> Sepal.Length  0.02312645  0.3728577 -0.03822814
#> Sepal.Width  -0.01905439  0.6122861  0.01530447
#> Petal.Length -0.01673156 -0.8337914  0.03709587
#> Petal.Width   0.04800159 -1.0873006 -0.04698136
#> 
#> , , B68
#> 
#>                    setosa versicolor     virginica
#> Sepal.Length  0.004920898  0.1849092 -0.0076034858
#> Sepal.Width   0.001708956  0.2397464  0.0154842171
#> Petal.Length -0.013521919 -0.3442411  0.0009316883
#> Petal.Width  -0.006786093 -0.3849375 -0.0335476884
#> 
#> , , B69
#> 
#>                   setosa versicolor     virginica
#> Sepal.Length  0.02856260  0.4047719 -0.0008024801
#> Sepal.Width   0.02475316  0.4662013  0.0070799115
#> Petal.Length -0.05034303 -0.7831053 -0.0033597015
#> Petal.Width  -0.03850551 -0.7783407 -0.0146636629
#> 
#> , , B70
#> 
#>                    setosa versicolor    virginica
#> Sepal.Length -0.003398841  0.2567616 -0.001801814
#> Sepal.Width  -0.007296735  0.3087283  0.014871242
#> Petal.Length  0.002666819 -0.4776974 -0.006111779
#> Petal.Width   0.008413800 -0.4976264 -0.030184895
#> 
#> , , B71
#> 
#>                 setosa versicolor virginica
#> Sepal.Length  1.544329   25.44272 -1.220838
#> Sepal.Width   1.673640   27.91585 -1.193190
#> Petal.Length -2.925344  -48.40866  2.231495
#> Petal.Width  -2.758600  -46.14450  1.916702
#> 
#> , , B72
#> 
#>                     setosa versicolor    virginica
#> Sepal.Length -0.0009954319  0.2173837  0.008663826
#> Sepal.Width  -0.0002488111  0.2362652  0.013784707
#> Petal.Length -0.0021090497 -0.3956916 -0.018417425
#> Petal.Width  -0.0027949162 -0.3759922 -0.023803128
#> 
#> , , B73
#> 
#>                 setosa versicolor virginica
#> Sepal.Length  1.788898   33.13725 -4.561541
#> Sepal.Width   1.944260   36.20001 -4.843388
#> Petal.Length -3.392094  -62.94980  8.578256
#> Petal.Width  -3.206798  -59.77786  7.944693
#> 
#> , , B74
#> 
#>                    setosa versicolor   virginica
#> Sepal.Length -0.005120978  0.4033044 -0.02779377
#> Sepal.Width  -0.040742395  0.6094947  0.01505030
#> Petal.Length  0.030623235 -0.8667828  0.02456492
#> Petal.Width   0.079831371 -1.0669829 -0.04209881
#> 
#> , , B75
#> 
#>                    setosa versicolor    virginica
#> Sepal.Length -0.005171119  0.2563015  0.008287958
#> Sepal.Width  -0.006328453  0.2860228  0.014326317
#> Petal.Length  0.008077434 -0.4805269 -0.018635319
#> Petal.Width   0.008908704 -0.4662183 -0.025339688
#> 
#> , , B76
#> 
#>                    setosa versicolor   virginica
#> Sepal.Length -0.007983701  0.2776987  0.00971989
#> Sepal.Width  -0.009293260  0.3072570  0.01420386
#> Petal.Length  0.014065307 -0.5232080 -0.02044949
#> Petal.Width   0.014354512 -0.5032607 -0.02461474
#> 
#> , , B77
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.03299768  0.5022889  0.003660013
#> Sepal.Width  -0.04097798  0.5698653  0.010384842
#> Petal.Length  0.06524796 -0.9650639 -0.010882361
#> Petal.Width   0.06912336 -0.9471812 -0.019540379
#> 
#> , , B78
#> 
#>                   setosa versicolor     virginica
#> Sepal.Length  0.02105162  0.4282125 -0.0008572707
#> Sepal.Width   0.01565418  0.4961474  0.0079265959
#> Petal.Length -0.03546112 -0.8307861 -0.0038740246
#> Petal.Width  -0.02309107 -0.8298135 -0.0164486947
#> 
#> , , B79
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.01857254  0.4985734 -0.007133922
#> Sepal.Width  -0.03637131  0.6158438  0.010748117
#> Petal.Length  0.04471509 -0.9886832  0.002068839
#> Petal.Width   0.06569858 -1.0417495 -0.024753327
#> 
#> , , B80
#> 
#>                    setosa  versicolor     virginica
#> Sepal.Length  0.009079881 0.072282171  0.0029305115
#> Sepal.Width   0.023558096 0.016181110  0.0039700200
#> Petal.Length -0.046125593 0.002426873 -0.0002690049
#> Petal.Width  -0.061012530 0.080660410 -0.0020739952
#> 
#> , , B81
#> 
#>                    setosa versicolor     virginica
#> Sepal.Length -0.004596970  0.2652375 -0.0008973241
#> Sepal.Width  -0.007816019  0.3130977  0.0145442747
#> Petal.Length  0.003781939 -0.4878040 -0.0069393955
#> Petal.Width   0.008414058 -0.5005383 -0.0290891177
#> 
#> , , B82
#> 
#>                    setosa versicolor     virginica
#> Sepal.Length  0.003775693  0.1778450 -0.0002600027
#> Sepal.Width   0.007084968  0.1922020  0.0127544944
#> Petal.Length -0.019840665 -0.2854818 -0.0054158831
#> Petal.Width  -0.021814811 -0.2742430 -0.0242095252
#> 
#> , , B83
#> 
#>                     setosa versicolor    virginica
#> Sepal.Length -4.023780e-04  0.2163059  0.004929772
#> Sepal.Width   3.008823e-05  0.2405853  0.013903973
#> Petal.Length -4.790065e-03 -0.3887922 -0.013624557
#> Petal.Width  -4.596176e-03 -0.3783545 -0.025339524
#> 
#> , , B84
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.09847426  0.01207895 -0.09274587
#> Sepal.Width   0.07157874  0.12301940 -0.02904584
#> Petal.Length -0.16453105 -0.09153986  0.13105392
#> Petal.Width  -0.10439977 -0.24510968  0.02035359
#> 
#> , , B85
#> 
#>                    setosa versicolor   virginica
#> Sepal.Length  0.057287852  0.2471260 -0.07327468
#> Sepal.Width  -0.003154084  0.5884954  0.02005583
#> Petal.Length -0.068117945 -0.6663322  0.07679000
#> Petal.Width   0.030038330 -1.0917903 -0.07127454
#> 
#> , , B86
#> 
#>                   setosa versicolor     virginica
#> Sepal.Length -0.02260010  0.4440881 -0.0059344058
#> Sepal.Width  -0.03895863  0.5533700  0.0126391589
#> Petal.Length  0.05093911 -0.8821606 -0.0005687662
#> Petal.Width   0.06903396 -0.9364819 -0.0280943833
#> 
#> , , B87
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.02788815  0.4404108  0.005602872
#> Sepal.Width  -0.03417636  0.4980675  0.011959741
#> Petal.Length  0.05463528 -0.8442682 -0.014215952
#> Petal.Width   0.05730588 -0.8264965 -0.021918935
#> 
#> , , B88
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.03220399  0.5186700  0.002482532
#> Sepal.Width  -0.04080784  0.5903771  0.009877887
#> Petal.Length  0.06379239 -0.9960446 -0.009078484
#> Petal.Width   0.06878962 -0.9805867 -0.018959673
#> 
#> , , B89
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.00492061  0.2883573 -0.009092893
#> Sepal.Width  -0.01653668  0.3821364  0.016120018
#> Petal.Length  0.01276411 -0.5735252  0.001520494
#> Petal.Width   0.02862253 -0.6432536 -0.036145081
#> 
#> , , B90
#> 
#>                   setosa versicolor     virginica
#> Sepal.Length -0.02129388  0.4370525 -0.0067343975
#> Sepal.Width  -0.03711044  0.5443079  0.0126894736
#> Petal.Length  0.04653876 -0.8604981  0.0005703995
#> Petal.Width   0.06444551 -0.9148269 -0.0283720905
#> 
#> , , B91
#> 
#>                    setosa versicolor   virginica
#> Sepal.Length  0.025009394  0.3120718 -0.06092793
#> Sepal.Width  -0.035527758  0.6476797  0.01944794
#> Petal.Length -0.009568739 -0.7781102  0.06224203
#> Petal.Width   0.081592424 -1.1815038 -0.06480119
#> 
#> , , B92
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.01945037  0.4606478 -0.011130257
#> Sepal.Width  -0.04104039  0.5955371  0.012379811
#> Petal.Length  0.04844062 -0.9287584  0.005922786
#> Petal.Width   0.07461133 -1.0153983 -0.029732636
#> 
#> , , B93
#> 
#>                    setosa versicolor    virginica
#> Sepal.Length -0.008103144  0.2893326  0.002733875
#> Sepal.Width  -0.011741068  0.3361665  0.014462621
#> Petal.Length  0.013148487 -0.5436769 -0.011736332
#> Petal.Width   0.017130765 -0.5474460 -0.027754021
#> 
#> , , B94
#> 
#>                    setosa  versicolor    virginica
#> Sepal.Length  0.006642918  0.14216297 -0.002888803
#> Sepal.Width   0.018080989  0.11843976  0.009554937
#> Petal.Length -0.041763949 -0.14293823  0.001781974
#> Petal.Width  -0.052583256 -0.09517398 -0.017169531
#> 
#> , , B95
#> 
#>                   setosa versicolor   virginica
#> Sepal.Length -0.01499694  0.4186296 -0.01521402
#> Sepal.Width  -0.03856090  0.5650978  0.01382463
#> Petal.Length  0.04029687 -0.8533314  0.01013037
#> Petal.Width   0.07045401 -0.9666765 -0.03426202
#> 
#> , , B96
#> 
#>                     setosa versicolor    virginica
#> Sepal.Length  0.0009842567  0.2449174 -0.014300279
#> Sepal.Width  -0.0113059734  0.3478841  0.017249762
#> Petal.Length  0.0019516215 -0.4974230  0.007272852
#> Petal.Width   0.0201547439 -0.5902212 -0.040517611
#> 
#> , , B97
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.01139087  0.3484754 -0.009597858
#> Sepal.Width  -0.02677691  0.4577614  0.015203831
#> Petal.Length  0.02793579 -0.6981776  0.002520907
#> Petal.Width   0.04749053 -0.7754494 -0.034647690
#> 
#> , , B98
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.01017972  0.3032064  0.004857091
#> Sepal.Width  -0.01417581  0.3500679  0.014452518
#> Petal.Length  0.01903322 -0.5776407 -0.014600430
#> Petal.Width   0.02273958 -0.5768805 -0.027080604
#> 
#> , , B99
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.01101733  0.05577544 0.001742916
#> Sepal.Width   0.04036975 -0.06674429 0.000364520
#> Petal.Length -0.07990856  0.16486311 0.005553394
#> Petal.Width  -0.11175853  0.31749167 0.006499457
#> 
#> , , B100
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length -0.01323664  0.3463279 -0.003364035
#> Sepal.Width  -0.02348086  0.4280205  0.014571085
#> Petal.Length  0.02781145 -0.6755713 -0.004622762
#> Petal.Width   0.03975810 -0.7146952 -0.030718122
#> 
#> , , B101
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.11728305 -0.05926954 -0.04597021
#> Sepal.Width  -0.06196319  0.01259188 -0.08822884
#> Petal.Length -0.10395970  0.06424119  0.11109232
#> Petal.Width   0.17505101 -0.05038429  0.16023812
#> 
#> , , B102
#> 
#>                    setosa  versicolor   virginica
#> Sepal.Length  0.057703589 -0.03545188 -0.11610425
#> Sepal.Width   0.003392407  0.02022119 -0.09390937
#> Petal.Length -0.072309470  0.03049599  0.19999690
#> Petal.Width   0.017234767 -0.05594363  0.14249474
#> 
#> , , B103
#> 
#>                    setosa   versicolor  virginica
#> Sepal.Length -0.004432472 -0.009100580 -0.1142798
#> Sepal.Width  -0.031306098  0.005346811 -0.1272109
#> Petal.Length  0.024968383  0.007729844  0.2185786
#> Petal.Width   0.061824633 -0.014678981  0.2109757
#> 
#> , , B104
#> 
#>                    setosa  versicolor   virginica
#> Sepal.Length  0.073938624 -0.05624089 -0.13888826
#> Sepal.Width   0.005029416  0.02658904 -0.08305977
#> Petal.Length -0.093094150  0.05180822  0.22092682
#> Petal.Width   0.020684515 -0.07758537  0.11089735
#> 
#> , , B105
#> 
#>                     setosa   versicolor   virginica
#> Sepal.Length  0.0302291151 -0.030360540 -0.09337563
#> Sepal.Width  -0.0602852732  0.008977546 -0.12285671
#> Petal.Length  0.0009121657  0.031326883  0.19042151
#> Petal.Width   0.1352535303 -0.030949920  0.21085637
#> 
#> , , B106
#> 
#>                   setosa   versicolor  virginica
#> Sepal.Length  0.01234514 -0.021162730 -0.0850532
#> Sepal.Width  -0.06816612  0.008941246 -0.1230981
#> Petal.Length  0.02759677  0.020158755  0.1804445
#> Petal.Width   0.14380606 -0.027031482  0.2148243
#> 
#> , , B107
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.09983797 -0.09071333 -0.19957390
#> Sepal.Width   0.05318573  0.09913306 -0.02827635
#> Petal.Length -0.15465512  0.04848665  0.26059723
#> Petal.Width  -0.06639036 -0.23946990 -0.02582549
#> 
#> , , B108
#> 
#>                   setosa   versicolor  virginica
#> Sepal.Length  0.03736473 -0.033499201 -0.1111537
#> Sepal.Width  -0.03356041  0.009798485 -0.1162724
#> Petal.Length -0.02447717  0.034632715  0.2079364
#> Petal.Width   0.08387986 -0.033931331  0.1900341
#> 
#> , , B109
#> 
#>                   setosa   versicolor  virginica
#> Sepal.Length  0.03255788 -0.031751189 -0.1086059
#> Sepal.Width  -0.04061907  0.008913418 -0.1198681
#> Petal.Length -0.01421317  0.033059023  0.2070870
#> Petal.Width   0.09622957 -0.031400688  0.1984144
#> 
#> , , B110
#> 
#>                   setosa   versicolor  virginica
#> Sepal.Length -0.02610518 -0.006719677 -0.1172145
#> Sepal.Width  -0.05640112  0.005212939 -0.1362581
#> Petal.Length  0.06702457  0.004916676  0.2278031
#> Petal.Width   0.10380593 -0.013411547  0.2281485
#> 
#> , , B111
#> 
#>                   setosa versicolor   virginica
#> Sepal.Length  0.08044094  0.1404434 -0.02882724
#> Sepal.Width   0.08040426  0.1748678 -0.02085126
#> Petal.Length -0.14812926 -0.2801542  0.04813322
#> Petal.Width  -0.12990568 -0.2969278  0.03038018
#> 
#> , , B112
#> 
#>                   setosa   versicolor   virginica
#> Sepal.Length  0.04022223 -0.003662322 -0.09537642
#> Sepal.Width   0.01976988  0.020715157 -0.09367698
#> Petal.Length -0.06128471 -0.008491504  0.17462544
#> Petal.Width  -0.02338759 -0.043661267  0.15068043
#> 
#> , , B113
#> 
#>                   setosa   versicolor   virginica
#> Sepal.Length  0.02652762  0.003896801 -0.09518517
#> Sepal.Width   0.01604568  0.016181739 -0.09983988
#> Petal.Length -0.04230092 -0.014856513  0.17823813
#> Petal.Width  -0.02154235 -0.031282754  0.16328915
#> 
#> , , B114
#> 
#>                   setosa   versicolor  virginica
#> Sepal.Length  0.03053137 -0.028246567 -0.1109258
#> Sepal.Width  -0.02966117  0.009223822 -0.1172925
#> Petal.Length -0.01858896  0.028600998  0.2083093
#> Petal.Width   0.07310235 -0.030567040  0.1922141
#> 
#> , , B115
#> 
#>                    setosa   versicolor  virginica
#> Sepal.Length -0.008797987 -0.014234393 -0.1047645
#> Sepal.Width  -0.062869153  0.007047888 -0.1318011
#> Petal.Length  0.050014187  0.012912176  0.2098726
#> Petal.Width   0.124198395 -0.020285078  0.2242902
#> 
#> , , B116
#> 
#>                    setosa   versicolor  virginica
#> Sepal.Length  0.013710257 -0.002998842 -0.1046449
#> Sepal.Width  -0.002083150  0.009774816 -0.1117885
#> Petal.Length -0.015372140 -0.002461807  0.1972178
#> Petal.Width   0.009973179 -0.021134322  0.1836379
#> 
#> , , B117
#> 
#>                   setosa   versicolor   virginica
#> Sepal.Length  0.07377167 -0.003481031 -0.09106159
#> Sepal.Width   0.04782616  0.051303175 -0.06429890
#> Petal.Length -0.11964362 -0.027828743  0.15101169
#> Petal.Width  -0.06642839 -0.105794178  0.09273325
#> 
#> , , B118
#> 
#>                    setosa   versicolor  virginica
#> Sepal.Length  0.020327217 -0.025532506 -0.1081809
#> Sepal.Width  -0.043575284  0.007712321 -0.1241679
#> Petal.Length  0.002511428  0.026243834  0.2092533
#> Petal.Width   0.097126446 -0.026358344  0.2073337
#> 
#> , , B119
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.07064394 -0.03011621 -0.00276732
#> Sepal.Width  -0.05359637  0.00933452 -0.07198728
#> Petal.Length -0.05244511  0.03080673  0.04837248
#> Petal.Width   0.13853897 -0.03157362  0.14526093
#> 
#> , , B120
#> 
#>                   setosa   versicolor   virginica
#> Sepal.Length  0.08350235  0.004395947 -0.08756006
#> Sepal.Width   0.05904843  0.071959464 -0.05151262
#> Petal.Length -0.13847231 -0.050294623  0.13879277
#> Petal.Width  -0.08516571 -0.144488664  0.06821801
#> 
#> , , B121
#> 
#>                    setosa   versicolor  virginica
#> Sepal.Length -0.009686367 -0.005416369 -0.1154722
#> Sepal.Width  -0.029009401  0.004835707 -0.1276857
#> Petal.Length  0.029924889  0.003566796  0.2203261
#> Petal.Width   0.054957254 -0.012099513  0.2114430
#> 
#> , , B122
#> 
#>                    setosa  versicolor   virginica
#> Sepal.Length  0.055111651 -0.02550514 -0.10834013
#> Sepal.Width   0.012565068  0.02240413 -0.09119886
#> Petal.Length -0.074886478  0.01703090  0.18886020
#> Petal.Width  -0.002501885 -0.05622455  0.14023070
#> 
#> , , B123
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.02542211 -0.02507928 -0.07449346
#> Sepal.Width  -0.06869132  0.00963606 -0.11781004
#> Petal.Length  0.01201423  0.02448967  0.16429031
#> Petal.Width   0.15034006 -0.03008170  0.20848231
#> 
#> , , B124
#> 
#>                   setosa versicolor   virginica
#> Sepal.Length  0.08115353  0.1372869 -0.02995081
#> Sepal.Width   0.08035663  0.1741054 -0.02074783
#> Petal.Length -0.14895879 -0.2758041  0.04945134
#> Petal.Width  -0.12950478 -0.2966691  0.02971297
#> 
#> , , B125
#> 
#>                    setosa   versicolor  virginica
#> Sepal.Length  0.033419385 -0.013744270 -0.1040957
#> Sepal.Width   0.002490740  0.014181258 -0.1039853
#> Petal.Length -0.042212503  0.007857605  0.1916705
#> Petal.Width   0.008907559 -0.034587857  0.1679957
#> 
#> , , B126
#> 
#>                   setosa   versicolor   virginica
#> Sepal.Length  0.06443333  0.001827536 -0.08703969
#> Sepal.Width   0.04316701  0.042110225 -0.07169270
#> Petal.Length -0.10537661 -0.028546430  0.15073113
#> Petal.Width  -0.06086159 -0.084882465  0.10944490
#> 
#> , , B127
#> 
#>                  setosa versicolor virginica
#> Sepal.Length -0.2421672 -0.2216412 -4.363708
#> Sepal.Width  -0.2706690 -0.2077871 -4.719176
#> Petal.Length  0.4638754  0.3997426  8.259781
#> Petal.Width   0.4493128  0.3301224  7.774650
#> 
#> , , B128
#> 
#>                   setosa versicolor  virginica
#> Sepal.Length  0.07120335  0.1465785 -0.1876002
#> Sepal.Width   0.06277482  0.2144841 -0.1788890
#> Petal.Length -0.12586587 -0.3122890  0.3401430
#> Petal.Width  -0.09790776 -0.3748599  0.2854783
#> 
#> , , B129
#> 
#>                   setosa   versicolor  virginica
#> Sepal.Length  0.01649826 -0.024551822 -0.1051267
#> Sepal.Width  -0.05139077  0.007586736 -0.1265160
#> Petal.Length  0.01205978  0.025128791  0.2070088
#> Petal.Width   0.11142511 -0.025693292  0.2133891
#> 
#> , , B130
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.08730936  0.08953597 -0.04734211
#> Sepal.Width   0.07881618  0.14632948 -0.02572547
#> Petal.Length -0.15549452 -0.20039315  0.07370162
#> Petal.Width  -0.12380586 -0.26017728  0.03255030
#> 
#> , , B131
#> 
#>                    setosa   versicolor  virginica
#> Sepal.Length  0.002259447 -0.009991824 -0.1123602
#> Sepal.Width  -0.025222336  0.006142117 -0.1236787
#> Petal.Length  0.013022902  0.008317090  0.2140348
#> Petal.Width   0.052247759 -0.016669031  0.2045938
#> 
#> , , B132
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.06897447  0.04191100 -0.06219727
#> Sepal.Width   0.06351301  0.06726268 -0.05593671
#> Petal.Length -0.12362610 -0.09304218  0.11065013
#> Petal.Width  -0.10034950 -0.11928798  0.08777784
#> 
#> , , B133
#> 
#>                    setosa   versicolor  virginica
#> Sepal.Length  0.005064428 -0.019776597 -0.1008618
#> Sepal.Width  -0.061071538  0.007574694 -0.1291438
#> Petal.Length  0.032022399  0.019326281  0.2034613
#> Petal.Width   0.126335093 -0.023672880  0.2205155
#> 
#> , , B134
#> 
#>                 setosa versicolor virginica
#> Sepal.Length -2.114890   4.976996 -37.24388
#> Sepal.Width  -2.303895   5.642571 -40.23084
#> Petal.Length  4.013580  -9.582897  70.46698
#> Petal.Width   3.802024  -9.396133  66.26035
#> 
#> , , B135
#> 
#>                   setosa versicolor    virginica
#> Sepal.Length  0.07533941 -0.2102184 -0.402309174
#> Sepal.Width   0.04634608  0.1356350  0.004792607
#> Petal.Length -0.12062288  0.1710119  0.486524205
#> Petal.Width  -0.06276048 -0.3637095 -0.177878929
#> 
#> , , B136
#> 
#>                   setosa    versicolor  virginica
#> Sepal.Length -0.03417959 -0.0004144675 -0.1239491
#> Sepal.Width  -0.04523566  0.0035142527 -0.1367795
#> Petal.Length  0.06986834 -0.0016928864  0.2363232
#> Petal.Width   0.07772125 -0.0073210928  0.2263942
#> 
#> , , B137
#> 
#>                   setosa   versicolor  virginica
#> Sepal.Length  0.01287440 -0.022829839 -0.1061037
#> Sepal.Width  -0.05114778  0.007307187 -0.1274189
#> Petal.Length  0.01631574  0.023208530  0.2087607
#> Petal.Width   0.10941518 -0.024404879  0.2148160
#> 
#> , , B138
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.08255999 -0.01415824 -0.10108948
#> Sepal.Width   0.05060015  0.06058785 -0.05722493
#> Petal.Length -0.13206984 -0.02063993  0.15879166
#> Petal.Width  -0.06839651 -0.12913939  0.07415513
#> 
#> , , B139
#> 
#>                  setosa versicolor virginica
#> Sepal.Length -0.5621278 -0.5184382 -8.931696
#> Sepal.Width  -0.6233762 -0.4974797 -9.654493
#> Petal.Length  1.0736811  0.9420215 16.903216
#> Petal.Width   1.0329623  0.7953292 15.903514
#> 
#> , , B140
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.05945198  0.04177363 -0.06280827
#> Sepal.Width   0.05890304  0.05480466 -0.06351625
#> Petal.Length -0.10915286 -0.08508307  0.11613816
#> Petal.Width  -0.09494906 -0.09400375  0.10294371
#> 
#> , , B141
#> 
#>                   setosa   versicolor  virginica
#> Sepal.Length -0.02690261 -0.005461984 -0.1193290
#> Sepal.Width  -0.05202025  0.004654897 -0.1362845
#> Petal.Length  0.06525738  0.003735173  0.2303939
#> Petal.Width   0.09456337 -0.011750957  0.2273198
#> 
#> , , B142
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.07008049  0.08296682 -0.04237445
#> Sepal.Width   0.07460760  0.09332051 -0.04438172
#> Petal.Length -0.13190431 -0.15927893  0.07931297
#> Petal.Width  -0.12244989 -0.15512184  0.07256558
#> 
#> , , B143
#> 
#>                    setosa  versicolor   virginica
#> Sepal.Length  0.057703589 -0.03545188 -0.11610425
#> Sepal.Width   0.003392407  0.02022119 -0.09390937
#> Petal.Length -0.072309470  0.03049599  0.19999690
#> Petal.Width   0.017234767 -0.05594363  0.14249474
#> 
#> , , B144
#> 
#>                   setosa   versicolor  virginica
#> Sepal.Length -0.00425071 -0.015919064 -0.1067075
#> Sepal.Width  -0.05859404  0.006760555 -0.1316157
#> Petal.Length  0.04180602  0.015141960  0.2121179
#> Petal.Width   0.11740153 -0.020404492  0.2230988
#> 
#> , , B145
#> 
#>                   setosa   versicolor  virginica
#> Sepal.Length -0.02770094 -0.006620960 -0.1161032
#> Sepal.Width  -0.05970268  0.005572247 -0.1363148
#> Petal.Length  0.07103094  0.004571830  0.2264872
#> Petal.Width   0.10985469 -0.014101180  0.2287291
#> 
#> , , B146
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.02954288  0.01493459 -0.08712369
#> Sepal.Width   0.02758060  0.02090442 -0.09301721
#> Petal.Length -0.05318137 -0.03123831  0.16416403
#> Petal.Width  -0.04374360 -0.03627448  0.15278158
#> 
#> , , B147
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.05189893  0.02218295 -0.07553982
#> Sepal.Width   0.04596574  0.03858308 -0.07514790
#> Petal.Length -0.09186380 -0.05110167  0.13891289
#> Petal.Width  -0.07178358 -0.06919527  0.12129009
#> 
#> , , B148
#> 
#>                   setosa  versicolor   virginica
#> Sepal.Length  0.06765893  0.04467715 -0.06067105
#> Sepal.Width   0.06380813  0.06678933 -0.05628651
#> Petal.Length -0.12219905 -0.09610131  0.10902556
#> Petal.Width  -0.10149060 -0.11716030  0.08913853
#> 
#> , , B149
#> 
#>                   setosa   versicolor  virginica
#> Sepal.Length  0.02407910 -0.018619001 -0.1093091
#> Sepal.Width  -0.01598312  0.009557503 -0.1144800
#> Petal.Length -0.01929804  0.016678851  0.2045752
#> Petal.Width   0.04257864 -0.027221414  0.1871625
#> 
#> , , B150
#> 
#>                   setosa   versicolor   virginica
#> Sepal.Length  0.09306556 -0.004749599 -0.09904084
#> Sepal.Width   0.06365612  0.091166509 -0.04100512
#> Petal.Length -0.15300258 -0.051181194  0.14617674
#> Petal.Width  -0.09055069 -0.187382277  0.04203626
#> 
#> attr(,"class")
#> [1] "Gradient" "array"   
#> attr(,"times_input")
#> [1] FALSE
#> attr(,"name")
#> [1] "Gradient"