LRP: Alpha-Beta-Rule for a dense layer

This is an alternative stabilizing method compared to the LRP-rule linear_simple_rule that does not leak relevance of treating negative and positive preactivation separately, i.e. we have a factor \(\alpha\) only for the positive part and another one \(\beta = 1 - \alpha\) for the negative part of the considered ratio between local and global preactivation: $$R_{ij} = ( \alpha \frac{z_{ij}^+}{z_j^+} + \beta \frac{z_{ij}^-}{z_j^-} ) \cdot R_j.$$ Then the relevance of the neuron \(i\) is given by the sum over the relevances of all possible outgoing connections, i.e. $$R_i = \sum_j R_{ij}.$$

linear_ab_rule(input, weight, bias, relevance, alpha = 0.5)

Arguments

input
The input vector with length dim_in of the hidden layer.

weight
The weight matrix with size (dim_in ,dim_out) of the hidden layer.

bias
The bias vector with length dim_out of the hidden layer.

relevance
A matrix with the relevance scores from each neuron of the upper hidden layer to each output neuron of the model.

alpha
Value of the factor for the positive contribution (default: \(0.5\)).

Value

Returns the relevance scores from each neuron in the lower hidden layer to each output neuron of the model as a matrix.

References

S. Bach et al. (2015) On pixel-wise explanations for non-linear classifier decisions by layer-wise relevance propagation. PLoS ONE 10, p. 1-46

See also