DeepLIFT: Reveal-Cancel-Rule

Implementation of the Reveal Cancel Rule introduced by Shrikumar et al. (2017) to determine a multiplier for an activation function. This rule defines different multipliers for the negative and positive contributions between difference-from-reference preactivation and postactivation. The positive and negative part of the postactivation \(\Delta y = \sigma(x) - \sigma(x')\) is given by $$\Delta y^+ := 0.5 ( \sigma(x' + \Delta x^+) - \sigma(x') + \sigma(x' + \Delta x^+ + \Delta x^-) - \sigma(x' + \Delta x^-) )$$ $$\Delta y^- := 0.5 ( \sigma(x' + \Delta x^-) - \sigma(x') + \sigma(x' + \Delta x^+ + \Delta x^-) - \sigma(x' + \Delta x^+) ).$$ Hence the multiplier are $$m_{\Delta x^+ \Delta y^+} = \frac{\Delta y^+}{\Delta x^+}$$ $$m_{\Delta x^- \Delta y^-} = \frac{\Delta y^-}{\Delta x^-}.$$ Afterward the contribution \(m_{\Delta x \Delta t}\) is calculated and returned.

reveal_cancel_rule(delta_x, mult_x_plus, mult_x_minus, layer)

Arguments

delta_x
Difference-from-reference preactivation of the current layer.

mult_x_plus
The multiplier from the upper positive difference-from-reference value to the output, i.e. \(m_{\Delta y^+ \Delta t}\)

mult_x_minus
The multiplier from the upper negative difference-from-reference value to the output, i.e. \(m_{\Delta y^- \Delta t}\).

layer
The hidden layer of type Dense_Layer.

Value

Returns the multiplier from the difference-from-reference preactivation to the output, i.e. \(m_{\Delta x \Delta t}\).

References

A. Shrikumar et al. (2017) Learning important features through propagating activation differences. ICML 2017, p. 4844-4866

See also